// api/model/invoiceStatus.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type InvoiceStatus = 'Pending' | 'InProgress' | 'Completed' | 'Unknown' | 'Error';

export const InvoiceStatus = {
    Pending: 'Pending' as InvoiceStatus,
    InProgress: 'InProgress' as InvoiceStatus,
    Completed: 'Completed' as InvoiceStatus,
    Unknown: 'Unknown' as InvoiceStatus,
    Error: 'Error' as InvoiceStatus
};