// api/model/roamingSessionStatus.ts
/**
 * AdminApp API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type RoamingSessionStatus = 'Rejected' | 'Undefined' | 'Requested' | 'Accepted' | 'Pending' | 'Charging' | 'Completed' | 'Faulted' | 'Cancelled';

export const RoamingSessionStatus = {
    Rejected: 'Rejected' as RoamingSessionStatus,
    Undefined: 'Undefined' as RoamingSessionStatus,
    Requested: 'Requested' as RoamingSessionStatus,
    Accepted: 'Accepted' as RoamingSessionStatus,
    Pending: 'Pending' as RoamingSessionStatus,
    Charging: 'Charging' as RoamingSessionStatus,
    Completed: 'Completed' as RoamingSessionStatus,
    Faulted: 'Faulted' as RoamingSessionStatus,
    Cancelled: 'Cancelled' as RoamingSessionStatus
};