// api/model/roamingTransactionDto.ts
/**
 * AdminApp API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TariffType } from './tariffType';
import { TariffIdleFee } from './tariffIdleFee';
import { RoamingSessionStatus } from './roamingSessionStatus';
import { UserBillingInfoDto } from './userBillingInfoDto';
import { TariffStartingFee } from './tariffStartingFee';
import { BasePriceType } from './basePriceType';
import { TariffMinuteFee } from './tariffMinuteFee';
import { TransactionStatus } from './transactionStatus';
import { ChargeSessionData } from './chargeSessionData';
import { RoamingInvoiceDto } from './roamingInvoiceDto';
import { TariffChargingFee } from './tariffChargingFee';
import { CurrencyDto } from './currencyDto';


export interface RoamingTransactionDto { 
    id?: string | null;
    chargePointInterestSessionId?: string | null;
    chargePointInterestSocketId?: string | null;
    chargePointInterestId?: string | null;
    currencyId?: string | null;
    externalTransactionId?: string | null;
    description?: string | null;
    currency?: CurrencyDto;
    readonly readableId?: string | null;
    receiptUrl?: string | null;
    paymentDate?: string | null;
    invoiceId?: string | null;
    creditCardNumber?: string | null;
    chargePointName?: string | null;
    roamingSessionStatus?: RoamingSessionStatus;
    paymentGatewayTransactionId?: string | null;
    taxRate?: number;
    totalFee?: number;
    totalTax?: number;
    totalKWH?: number;
    totalMinutes?: number;
    userId?: string | null;
    taxNo?: string | null;
    taxAdministration?: string | null;
    taxName?: string | null;
    totalAmount?: number;
    discountAmount?: number | null;
    tariffType?: TariffType;
    tariffBasePrice?: number;
    tariffBasePriceType?: BasePriceType;
    chargeSessionData?: ChargeSessionData;
    tariffMinuteFee?: TariffMinuteFee;
    tariffChargingFee?: TariffChargingFee;
    tariffIdleFee?: TariffIdleFee;
    tariffStartingFee?: TariffStartingFee;
    plateNo?: string | null;
    readonly roamingSessionId?: string | null;
    connectorId?: string | null;
    chargePointId?: string | null;
    paymentMethodId?: string | null;
    created?: string;
    status?: TransactionStatus;
    isInvoiceDownloadable?: boolean;
    memberBillingInfo?: UserBillingInfoDto;
    invoice?: RoamingInvoiceDto;
    metadata?: string | null;
}
export namespace RoamingTransactionDto {
}