// api/model/tariffType.ts
/**
 * AdminApp API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type TariffType = 'Undefined' | 'PublicTariff' | 'UserTariff';

export const TariffType = {
    Undefined: 'Undefined' as TariffType,
    PublicTariff: 'PublicTariff' as TariffType,
    UserTariff: 'UserTariff' as TariffType
};