// api/model/transactionStatus.ts
/**
 * AdminApp API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type TransactionStatus = 'Refunded' | 'Failed' | 'Undefined' | 'Pending' | 'WaitingForPayment' | 'WaitingForPaymentApproval' | 'Success' | 'SuccessWithoutPayment' | 'CompletedManual';

export const TransactionStatus = {
    Refunded: 'Refunded' as TransactionStatus,
    Failed: 'Failed' as TransactionStatus,
    Undefined: 'Undefined' as TransactionStatus,
    Pending: 'Pending' as TransactionStatus,
    WaitingForPayment: 'WaitingForPayment' as TransactionStatus,
    WaitingForPaymentApproval: 'WaitingForPaymentApproval' as TransactionStatus,
    Success: 'Success' as TransactionStatus,
    SuccessWithoutPayment: 'SuccessWithoutPayment' as TransactionStatus,
    CompletedManual: 'CompletedManual' as TransactionStatus
};